cluster_chart/requirements.txt
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
pandas==2.2.2
plotly==5.21.0
python-dotenv==1.0.1